fix: RoomController 분리
